Struct rollsum::bup::Bup

pub struct Bup {
    // some fields omitted
}

Rolling checksum method used by bup

Strongly based on https://github.com/bup/bup/blob/706e8d273/lib/bup/bupsplit.c https://github.com/bup/bup/blob/706e8d273/lib/bup/bupsplit.h (a bit like https://godoc.org/camlistore.org/pkg/rollsum)

Methods

impl Bup

fn new() -> Self

Create new Bup engine

fn find_chunk_edge(&mut self, buf: &[u8]) -> Option<usize>

Find chunk edge using Bup defaults.

See Engine::find_chunk_edge_cond.

fn count_bits(&self) -> u32

Counts the number of low bits set in the rollsum, assuming the digest has the bottom CHUNK_BITS bits set to 1 (i.e. assuming a digest at a default bup chunk edge, as returned by find_chunk_edge). Be aware that there's a deliberate 'bug' in this function in order to match expected return values from other bupsplit implementations.